refactor(script): separate validation logic from UI
