+++ script/src/validate-monorepo/validation/circleci.ts
